Live Edit Service
Initial implementation of live edit service.

The goal of this CL is mostly checking in the prototype's code and
get the team to start testing and working on live edit.

My guess is LiveEditService (or what becomes of it) will need to be
merged / rewrittend to server a similar purpose for Preview so we can
view this only a temporary solution for now.

Bug: None
Test: None

Change-Id: I94dab35b9a9df126d215f45f72bdb05ebb17b653