Fix name conflict resolution in Animation Preview tabs
We were always displaying a number (e.g. #1, #2) next to the tab name in the
Animation Preview to help users identify tabs that share the name with other
tabs. The UX recommendation is to not show a number for the first tab (as
the user will rarely use the same label for different tabs on purpose) and
use numbers only in the tabs with repeated name. The format was also changed
from something more similar to file naming in operational systems.

Before this change, we'd have something like:

* tabName #1, anotherTab #1, tabName #2

After this change, we will have something like:

* tabName, anotherTab, tabName (1)

Bug: N/A
Test: ComposePreviewAnimationManagerTest and Manual/Visual
Change-Id: I7aa414cca23a576c6e6bc9ffcc6f254685eecba3