APK Sign: reload passwords on keystore or key change
Listen for the change is keystore and key selection to reload passwords
for the new selected pair.

Keystore can be selected using file dialog or typing the path directly
in the field. To support both we trigger password reloading both
in the end of file selector dialog process and on listener action of the
textfield itself (this one triggers when user clicks enter after text
input). Keystore also can be created using a different dialog, don't do
anything in this case as newly created passwords are already
prepopulated here.

Similar logic applies to the key itself: it can be created/selected
through the dialog (though it is the same single dialog in this case)
or alias could be typed manually and then we wait user to hit enter
to try load the password.

Before loading the password we clear both password fields on keystore
selection change and only key password on key selection change as
passwords left in the fields on these events are most likely to be
incorrect (from a different key) and user is expected to type a new
ones for new selections (if they were not saved previously).

Bug: 191306261
Test: Added test case
Change-Id: I96b2fa4d0585d4881573d19930fa4f3f07242735