[Dsl] Reorganize inheritance of interfaces
We have a need for a more generic way of getting from a Dsl model to a
corresponding Dsl element and its holder.  Define a new interface,
GradleDslElementModel, to capture that functionality.

However, this functionality is not implementable for GradleFileModel
objects, as the corresponding Dsl element (a GradleDslFile) has no
holder.  This means we need to distinguish between blocks and files,
which otherwise look quite similar (each of them effectively being a
container for properties, with files having additional fields and
methods on top.  Therefore, define a new GradleBlockModel and use that
in the definitions of all the individual typed block models rather than
GradleDslModel, which remains a superinterface.

We also now need to distinguish between a PluginsModel, which need not
have an associated PsiElement (because of other ways of applying plugins
in build.gradle files, such as `apply plugin:`) and a PluginsBlockModel,
which is a representation of a plugins block (for example in the context
of a `pluginManagement` block in settings.gradle).

This reorganization also allows us to remove the slightly weird
implementation of delete() on GradleFileModel, which was there only to
satisfy a similar issue of inheritance where because something needed to
be implemented on GradleDslModel it necessarily needed to be implemented
on GradleFileModel.

Bug: n/a
Test: existing
Change-Id: Ibe3c190772d2ce1affaaca188aa66cc01ef022ef