Do not call Configuration#getTheme in EDT
This is done to detect configuration changes in LayoutlibSceneManager
but it can lead to slow operations. This moves the change detection to
using the listener instead of the manual detection.

This CL also removes the bulk notification when locale, device or target
change at the project level since the configurations are correctly
notified about the individual events.

Bug: 193700209
Test: Existing test pass
Change-Id: If537c0c0c097ae8ca982452ab4031faaa3a6de51