Make Transport Pipeline Demo show debuggable processes only
Transport pipeline recently adds the support of profileable
processes. As a result, the pipeline's PROCESS data include
both debuggable and profileable processes.

TransportPipelineDialog aims to demo the full capability of
the pipeline, of which the JVMTI agent is an important
functionality. As JMVTI agent is supported by debuggable
processes only, we show them only and ignore profileable
ones.

Bug: n/a
Test: manual
Change-Id: I2384cf052221d06ea80d80e0b7aa67fdfae2b452