[Dsl] Make testInstrumentationRunnerArguments a MUTABLE_MAP
Also, make the method of the same name have (new) AUGMENT_MAP semantics,
implemented by calling the newly-defined addToParsedElementMap() helper,
allowing us to delete the special-case implementation.  We continue to
need a special-case implementation of the singular
`testInstrumentationRunnerArgument`, which has semantics not seen
elsewhere (so it's not entirely clear that making a general
implementation would be a win).

Bug: 194192720
Test: existing
Change-Id: Ia9965c0c5087bf11d87b9275c8a5321c3ee8a72a