Check section row by its tree node value
Before this commit, we use tree.getPathForRow() to check if a row
is immediate children of the root, which is not updated
when rendering component. This commit uses TreeNode value from
model which is guaranteed to be consistent.

Bug: 197250562
Test: Manual Test
Change-Id: I80b28701bb92198d7c043ff35bf4e381e905e28b
(cherry picked from commit 30963308eaecfdb74fc6a18b3f6a92c2ac7fd84a)