[AUA] Check for forced upgrade check in terms of new rules
This effectively removes special-case handling of 2.4.0-alpha08, which
apparently was so disruptive that projects of earlier 2.4 alphas were
supported perpetually, with no requirement to upgrade.  (This was
probably not actually true in the past, and is definitely theoretical
now given that the earliest supported version of AGP is around 3.1 or
3.2).  Also, we now no longer force an upgrade from a preview to a later
snapshot, so adjust the test checking for that to include a snapshot in
the possible versions.

Bug: 189388326
Test: existing tests adjusted
Change-Id: Ia7442f4b05d03f6c3737707a063c7dda0fafecad