[AUA] Fix preservation of selected item across changes
When we receive a list of known or suggested versions, we need to
preserve the current selection state.  Make sure to read and preserve
that state before modifying the elements of the control, and restore it
afterwards.

This changes no behaviour in the current implementation, because the
first added element was in all cases the latest known version, but with
an upcoming change to pass in other recommended versions, the effects of
this change become visible.

Bug: 189388326
Test: existing
Change-Id: I4ac837cd6740504b2121371dcc25ce9947700d2d