Use only the holder modules for compose preview selection
In the future the number of modules with the Android Facet
on them will increase to also include unit test and Android test
modules. This change adjusts the module selection combobox to filter out
these modules. e.g MyApp.app will be displayed instead of
MyApp.app.main, MyApp.app.unitTest and MyApp.app.androidTest

It is worth noting that these modules will no longer contain any source
or deps and if these are required (they aren't in this case) then the
correct module must be selected via the methods in ModuleUtil.kt

This change should not affect current behaviour

Bug: N/A
Test: current behaviour unaffected
Change-Id: I81a81e6826ffac8f64d93846871af44cc9f2a9ad