Attach long field values as files instead
Crash has a limit on the number of bytes for fields in the request body. This change sanitizes fields to ensure Crash doesn't exclude information from reports. Fields that are too long for Crash are truncated and the original raw text is included as a file.

For the future, we intend to write a wrapper around MultipartEntityBuilder to abstract this logic and pass around this object in the serialization methods for reports.


Change-Id: I9ec80bc56d6255f96e3bb6776254508125d10aeb
Bug: 196853911