Fix ResourceIdProvider inconsistency
The ResourceClassGenerator relies on an internal cache of id that can be
generated by the ResourceIdProvider. The ResourceIdProvider can be
invalidated independently of the ResourceClassGenerator by calling
ResourceIdProvider.resetDynamicIds. Doing this without clearing the
ResourceClassGenerator can lead to inconsistencies on the id usages.

This CL introduces the concept of generation so the users of
ResourceIdProvider can invalidate their own copies of the ids when
needed.

Test: Added test for generations of id
Bug: 197931929
Change-Id: I17c78675633ae41ec89b6c224883285c1336cac9
(cherry picked from commit 786e2e0ba22b80ccb3b5f0dae4318764a1a5570b)