System images with extensions placed in relevant tab
In AVD manager, system images that contain SDK extensions
but are not the base for that API level are placed in
the OTHER or X86/ARM tab.

The center tab has been renamed to ARM Images when studio
is run on an M1 machine. The center column now refers to
performant images that are not recommended.

Test: ChooseSystemImagePanelTest
Bug: 198338419
Change-Id: Ib490c08f95954dec627929f42150ac6c04b1b2eb
(cherry picked from commit 112075e9b2da7dfa50e15d7cc28457434bd5be28)