Put SelectNextColumnCellAction in the action map
It focuses the next cell or AvdActionPanel component when you type the
Tab key. At the end of the row it focuses the first cell in the next
row. At the end of the table it focuses the first cell in the first row.

Bug: 159782511
Test: Not applicable
Change-Id: Ia303a01b7509f8425e0858ae1049801ebee0215f
(cherry picked from commit a2732d1b2c4097711fd273f0e0f78dade5d74094)