Partially revert 89049e11535
89049e11535be1e63e341e0cf9a32ef7d3d01366 broke the way the details panel
worked with the keyboard. We want it to be driven by changes in the
table selection, regardless if the selection was changed with the mouse
or keyboard.

As the right click menu in the Virtual table is still gone, this is a
good enough mitigation for Issue 200134983

(cherry picked from commit 6936d940f6f2dfd17961d4c2029ccddf07e9afd0)

Bug: 200134983
Test: Not applicable
Change-Id: I54d6e6880c3a92d5c062d7e087b21a7ead2c9cfa