Device manager: fix NPE when opening device explorer
When opening the Device File Explorer for the first time *and*
when the new toolwindow shows up in the same pane as Device Manager
(the default behavior), the IJ platform calls "uninstallUI/installUI"
on all components. This creates a problem in BasicTableUI, which ends
up being called on the "mouseRelease" event with a `null` JTable field.

The "fix" is to open the Device File Explorer toolwindow after all enqueued
mouse events have been processed, which is achieved by using "invokeLater".

Here is the offending code in the IJ platform:
https://github.com/JetBrains/intellij-community/blame/a98539af08c596b3da3b97906d54db0af3511236/platform/platform-impl/src/com/intellij/openapi/wm/impl/ToolWindowsPane.java#L856

Fixes: 200165926
Test: Existing
Change-Id: I3a45292ac9451e473a606dced87e31ee7d51e0a7
(cherry picked from commit 006bdf7e9bbb1090dc2666de90e8553650b6362c)