Avoid image already disposed exception
This change removes the assertion when the image was disposed and moves
it to a warning. This will avoid triggering a user visible error on
"debug" builds. The assert was harmless in most cases but it would
trigger a visible user error.
This change also addresses the root cause that was causing the
exceptions, when an image was still being used and it was disposed in a
separate thread.

The dispose responsibility has been moved to a new ImagePoolImageDisposer
class that is the only way to "ask" for image to be removed. This is
just a request, since ImagePool images do not need to be manually
disposed and the only reason to do so is to improve performance by
forcing disposals to happen before the GC kicks in.
The new ImagePoolImageDisposer also allows to have critical sections
where an image will not be disposed. This allows, for example, for a
.paint() method in a JComponent to ensure that a pooled image is not
going to be disposed mid paint. Once the critical sections over the
image end, the image will be disposed immediately.

Test: Adds new tests for ImagePoolImageDisposer.
Fixes: 183031601
Change-Id: I8da05ee26f3f329de58d86400000ab7ce9edb01c