Disable logging unit tests execution on the IDE side.
in
https://studio-stats.corp.google.com/#report_id=Testing/Test%20runs%20total
we get unit test execution logged twice : this is because we do log the
test execution on the IDE side, and again on the AGP side in order to
get the test libraries being used.

Logging tests execution on the IDE side is not providing any additional
information than the AGP's, so it makes more sense to log events from
the AGP side only.

Bug: N/A
Test: on AGP side
Change-Id: Iaeeb195e7febd52557e293f411f1ad7e62ebea7c