Extract Native Code Navigation From IntelliJ Code Navigation
To help isolate the different methods of navigating to different source
files, this change moves the native code navigation out of the large
IntelliJ code navigation type.

This leaves the (previously large) IntelliJ with only one form of code
navigation which will be addressed in a larger CL based on some of the
TODOs left in the code over the last couple changes.

Bug: 197112399
Test: Manually test using native code Android Studio project
Change-Id: Ic788c0dc379ed51573fad171b6d9ca38702c785a