Fix incorrect editor warnings from overlayables
Overlayable names do not appear in the R class, so AGP doesn't need to
validate that the name is a valid java identifier.

Change I08e6d747a3b1f46dffc026e49160a189a65cfb75 relaxed the name
validation for overlayables, but the type was not unconditionally passed
down when validating in the editor.

Also fixes the Overlayable.Policy.type defition to be a
GenericAttributeValue rather than an AndroidAttributeValue.

Bug: 201525574
Test: Added AndroidValueResourcesTest.testOverlayable
Change-Id: I9e166ffca677733f5144cff2668c9d96e1973382