Correctly display grouped issues in Visual Linting
The VisualLintRenderIssue object cannot use hashcode and equals
inherited from Issue, as its description can change. So instead only
have equality between the same exact objects, and use a hascode that
only takes into account non-changing fields.
In addition, because the description of the issue is not fixed, allow
for the IssueView to be updatable to reflect a possible change of
description.

Bug: 196596237
Test: N/A
Change-Id: I5d1002881a2d05d74287db4627966f7d507d17b1