[Dsl,.toml] Better handling of psiToName on keys
`TomlKey`.`name` returns null if the key is made of more than one
segment, which is exactly the case that we should be trying to handle
here.  Instead, somewhat uglify the code to join the names of segments,
except that if any segment has no name (not sure when that could
happen, but it's in the types) simply escape the key text.

(At the moment the visitor itself manages access to key segments, but
it's not impossible that something will lead that to change in the
future.)

Bug: 200280395
Test: existing
Change-Id: I29d90bb3c7d9f11413873e41a0616c7eb8c5fca2