[Dsl] Prefer computing internal names from Psi
Rather than converting the Psi to a String and then resolving the
reference from text, which loses information about how text is
interpreted in the external language, attempt to compute the internal
name directly from the Psi.  Provide an interface for converting an
element to internal text, and use it where appropriate, including when
generating a name for an injection, which now rather than being a String
in some arbitrary syntax can be the internal (GradleNameElement) name of
the target.

Bug: 142454204
Test: existing
Change-Id: I709f49e0b2245801987fa5715f18d7d3eba12767