Display compile error in the Build tool window
This patch fixes two issues in GradleConnectedAndroidTestInvoker.

1. A wrong ProjectSystemId was used in the externalTaskId.

The ProjectSystemId should have been GradleConstants.SYSTEM_ID
so that output messages are parsed and translated into
build events correctly.

2. Display Build tool window upon build failure except for test failures.

GradleBuildInvokerImpl has a native support for showing the
Build tool window upon build failures however we cannot use
it here because both compile error and test failure are
considered as a build failure. We don't want to show the Build
tool window when it is a test failure. So, in this patch,
it checks if the build fails before a test suite starts and
if so, it shows the Build tool window. Gradle plugin has a
similar implementation in
https://github.com/JetBrains/intellij-community/commit/1618f0c71551e51b222e5e24738f826b3a6b47a6

Fixes: 201562459
Test: GradleConnectedAndroidTestInvokerTest and tested manually
Change-Id: Idb9e1c28507fb8363a3861920dc721cf7ea1c6a1