[AUA] Move package from manifest to build Dsl
In each (Gradle) project, inspect as many of the manifests as we can
reasonably find, in order to intuit the namespace that the manifests
declare in their package attribute, and to delete those attributes.  The
package declaration from a manifest in a subdirectory named "main"
supersedes other declarations, though that in turn is superseded by an
already-existing namespace declaration in the Dsl.

Bug: 200682321
Test: added, existing integration test adjusted
Change-Id: I8149345854488dfcd79aaf00f14d8bc9bfcee208