[AUA] Upgrade variable version declarations
In pluginManagement->plugins blocks, the syntax of the `id` operator is
less constrained, not requiring the version to be a literal constant.
Remove the defensive code that only upgrades the declaration if it is a
literal string.  This will have no effect on the upgrade of legal
version declarations in plugins blocks in other contexts, where the
requirement of a literal constant is retained; semantically-illegal
declarations will be upgraded to equally-illegal ones where previously
they would have been left alone.

Bug: 201946668
Test: added
Change-Id: I80211a00791a93621a7019d5113d3fbd716a9ade