Add Common Issue Panel into Problems Panel
Use new common issue panel instead of old design surface panel.
Click the issue button in layout editor and/or validation tool to
attach the issues in common issue panel.

At this moment, the issue shows in common issue panel is html style,
it needs to be updated.

The ui and content of tree nodes are still in development, the tests
of nodes will be added after that.

Bug: 194780029
Test: manually.
Change-Id: Ibcd583459514594acc05a2dc07ee5ca6c11fefbd