Add Jetbrains namespace to ui-tooling-preview resolution
When deciding whether we should look for ui-tooling implementation files,
e.g. ComposeViewAdapter, in ui-tooling or ui-tooling-preview, we're checking
if the project has a dependency to androidx.compose.ui:ui-tooling-preview.
However, Jetbrains also publish this artifact but with a different namespace
("org.jetbrains.compose").

Jetbrains reported that Compose Preview is not working as expected when they
use their own artifacts, and the error message gives a hint that we're
looking for classes in the wrong place: "The following classes could not be
found: - androidx.compose.ui.tooling.preview.ComposeViewAdapter".

In order to fix this issue, we should include the Jetbrains namespace when
resolving the dependency to the ui-tooling-preview artifact.

Bug: 202135291
Test: N/A
Change-Id: I7181fabe5e5e18c2f080915ffcfeec71be9af18b
(cherry picked from commit 4a54aadf250848c7b7da142b159218efecb15b87)