Fix PackageClassConverter manifest package resolution
When using the namespace DSL, the package attribute will
not be set in AndroidManifest.xml. This means for Activity
tags that are allowed to use the short name of the class only,
we must take the module package from AndroidManifestUtils,
which uses the AndroidModuleSystem.

Test: GradleBuildFileNamespaceRClassesTest.kt
Fixes: 202006729
Change-Id: I87dddf83101a821d7ae2d7b3a5c03aa306ae1191
(cherry picked from commit a3c68962d4bbe08cdda8293180dbc6f33e0dd253)