Add support for function literals to LiveEdit
Updates the studio generation code to pass "support classes" to
LiveEdit. The set of support classes is currently the set of classes
built when compiling a kotlin file, excluding the class built from the
file itself.

Adds additional handling to support anonymous inner classes to the
name resolution code within the LiveEditCodeGenerator.

Bug: None
Test: Manual
Change-Id: I2aad3d4644abceca325ea3baef906eab7cf4d08f