Remove old code path for ATF
We no longer use "unoptimized" version of ATF integration path. Removing
the code to clean up the code.

This also update all tests (including perfgate test) to better
reflect the current code path.

With this cl I expect perfgate atf integration to go down drastically
and we can use it as a better benchmark when new ATF library is added.

Bug: 181892299
Test: Manally tested
Change-Id: I916e9083aeede5136414e1d0de7114c944acf5a5