Optimize AndroidMissingOnClickHandlerInspection
Previously, running this inspection on any layouts would
take ~ 40-350ms. This is regardless of whether the
layout has an onClick attribute.

This change pushes the expensive findRelatedActivities
code to after an onClick attribute is found.

This change also reduces the amount of isInheritor checks
needed as the result of the AndroidGotoRelatedLineMarkerProvider
already has Context checks for activities and fragments.

As a result , here are the times to run the inspection now:
* A layout without onClick attribute: 19ms
* A layout with onClick attribute, but also tools:context set: 19ms
* A layout with onClick attribute, but no tools:context set: 27ms

Test: Existing, functionality has not changed
Bug: 202744573
Change-Id: I5037f82a4c5e7222b77ac32c5d8f00f04f25c902