[MPSS] Allow DynamicFeatureParameters to work correctly with mpss
This change adjusts a few methods to allow them to work with mpss or be
more explicit about which modules are being used and passed around.

The main part of this change adjusts getFeatureModulesDependingOnFeature
and getFeatureModuleDependenciesForFeature to use the main module (with
mpss this is the module with the dependency information). We also filter
the modules to remove all modules that belong to the same Gradle
project, this prevents the method from thinking they depend on
themselves.

Bug: N/A
Test: N/A
Change-Id: I6b465f108ad81585043e7d14217cf8eca2e84228